feat: id 타입 Long으로 변환

+++ src/main/java/kr/codesqaud/cafe/cafeservice/domain/Article.java
@@ -1,10 +1,12 @@
 package kr.codesqaud.cafe.cafeservice.domain;
 
 
+import jdk.jshell.Snippet;
+
 import java.time.LocalDateTime;
 
 public class Article {
-    private int id;
+    private Long id;
     private String writer;
     private String title;
     private String content;
@@ -23,11 +25,11 @@ public class Article {
     public Article() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
